+++ src/language/Keywords.kt
@@ -3,5 +3,5 @@ package language
 val KEYWORDS = arrayOf(
         "set",
         "tell"
-        //TODO: Add more
+        //TODO: Add Java keywords, or change any variable name that is a Java keyword
 )