+++ src/language/objects/LivingEntity.kt
@@ -2,9 +2,9 @@ package language.objects
 
 import language.DataType
 
-class LivingEntity(name: String, mutable: Boolean = false): VulcanObject(DataType.ENTITY, name, mutable) {
+class LivingEntity(name: String, java: String = name, mutable: Boolean = false): VulcanObject(DataType.ENTITY, name, java, mutable) {
 
-    override val validMessages: Map<String, Int> = mapOf(
+    override val actions: Map<String, Int> = mapOf(
             Pair("jump", 0),        //Make entity jump
             Pair("take", 2),        //Damage entity
             Pair("burn", 3),        //Set entity on fire
@@ -20,13 +20,13 @@ class LivingEntity(name: String, mutable: Boolean = false): VulcanObject(DataTyp
     override fun convertMessage(message: String, parameters: Array<String>, variables: Map<String, VulcanObject>): String {
         when(message) {
             //Jump
-            "jump" -> return "MessageUtils.makeJump($name);"
+            "jump" -> return "MessageUtils.makeJump($java);"
 
             //Damage
             "take" -> {
                 if(parameters[1] == "damage") {
                     val damage = DataType.INTEGER.toJava(parameters[0], variables)
-                    return "MessageUtils.attack($name, $damage);"
+                    return "MessageUtils.attack($java, $damage);"
                 } else {
                     throw IllegalArgumentException("invalid syntax")
                 }
@@ -36,7 +36,7 @@ class LivingEntity(name: String, mutable: Boolean = false): VulcanObject(DataTyp
             "burn" -> {
                 if(parameters[0] == "for" && (parameters[2] == "seconds" || parameters[2] == "second")) {
                     val time = DataType.INTEGER.toJava(parameters[1], variables)
-                    return "$name.setFire($time);"
+                    return "$java.setFire($time);"
                 } else {
                     throw IllegalArgumentException("invalid syntax")
                 }
@@ -46,20 +46,20 @@ class LivingEntity(name: String, mutable: Boolean = false): VulcanObject(DataTyp
             "teleport" -> {
                 if(parameters[0] == "to") {
                     val position = DataType.VECTOR3.toJava(parameters[1], variables)
-                    return "$name.setPosition($position.getX(), $position.getY(), $position.getZ())"
+                    return "$java.setPosition($position.getX(), $position.getY(), $position.getZ())"
                 } else {
                     throw IllegalArgumentException("invalid syntax")
                 }
             }
 
             //Restore air
-            "breathe" -> return "$name.setAir(300);"
+            "breathe" -> return "$java.setAir(300);"
 
             //Swing arm
             "swing" -> {
                 if((parameters[0] == "left" || parameters[0] == "right") && parameters[1] == "arm") {
                     val mainHand = parameters[0] == "right"
-                    return "MessageUtils.swingArm($name, $mainHand);"
+                    return "MessageUtils.swingArm($java, $mainHand);"
                 } else {
                     throw IllegalArgumentException("invalid syntax")
                 }
@@ -69,26 +69,26 @@ class LivingEntity(name: String, mutable: Boolean = false): VulcanObject(DataTyp
             "heal" -> {
                 if(parameters[1] == "health") {
                     val amount = DataType.INTEGER.toJava(parameters[0], variables)
-                    return "$name.heal($amount);"
+                    return "$java.heal($amount);"
                 } else {
                     throw IllegalArgumentException("invalid syntax")
                 }
             }
 
             //Instant kill
-            "die" -> return "MessageUtils.kill($name);"
+            "die" -> return "MessageUtils.kill($java);"
 
             //Set riding entity
             "ride" -> {
                 val target = DataType.ENTITY.toJava(parameters[0], variables)
-                return "$name.startRiding($target, true);"
+                return "$java.startRiding($target, true);"
             }
 
             //Explode the entity (KABOOM!)
             "explode" -> {
                 if(parameters[0] == "with" && parameters[1] == "strength") {
                     val strength = DataType.FLOAT.toJava(parameters[2], variables)
-                    return "MessageUtils.explode($name, $strength);"
+                    return "MessageUtils.explode($java, $strength);"
                 }
 
                 throw IllegalArgumentException("invalid syntax")