+++ src/language/objects/VulcanBoolean.kt
@@ -2,15 +2,15 @@ package language.objects
 
 import language.DataType
 
-class VulcanBoolean(name: String, mutable: Boolean = false): VulcanObject(DataType.BOOLEAN, name, mutable) {
+class VulcanBoolean(name: String, java: String = name, mutable: Boolean = false): VulcanObject(DataType.BOOLEAN, name, java, mutable) {
 
-    override val validMessages: Map<String, Int> = mapOf(
+    override val actions: Map<String, Int> = mapOf(
             Pair("flip", 0)     //Flip boolean state
     )
 
     override fun convertMessage(message: String, parameters: Array<String>, variables: Map<String, VulcanObject>): String {
         when(message) {
-            "flip" -> return "$name = !$name;"
+            "flip" -> return "$java = !$java;"
         }
 
         throw IllegalArgumentException("unsupported message $message")