+++ src/language/objects/VulcanObject.kt
@@ -2,17 +2,25 @@ package language.objects
 
 import language.DataType
 
-abstract class VulcanObject(val type: DataType, val name: String, val mutable: Boolean = false) {
+abstract class VulcanObject(val type: DataType, val name: String, val java: String = name, val mutable: Boolean = false,
+                            vararg fieldObjects: VulcanObject) {
 
-    abstract val validMessages: Map<String, Int>
+    val fields: Map<String, VulcanObject>
+    abstract val actions: Map<String, Int>
 
-    fun isValidMessage(message: String): Boolean = message in validMessages.keys
+    init {
+        val mutableFieldMap: HashMap<String, VulcanObject> = hashMapOf()
+        fieldObjects.asSequence().forEach { mutableFieldMap[it.name] = it }
+        fields = mutableFieldMap.toMap()
+    }
+
+    fun isValidMessage(message: String): Boolean = message in actions.keys
 
     protected abstract fun convertMessage(message: String, parameters: Array<String>, variables: Map<String, VulcanObject>): String
 
     fun messageToJava(message: String, parameters: Array<String>, others: Map<String, VulcanObject>): String {
         if(!isValidMessage(message)) { throw IllegalArgumentException("unrecognised message $message") }
-        if(parameters.size != validMessages[message]) { throw IllegalArgumentException("improper usage of $message") }
+        if(parameters.size != actions[message]) { throw IllegalArgumentException("improper usage of $message") }
         return convertMessage(message, parameters, others)
     }
 }