+++ src/language/objects/World.kt
@@ -2,9 +2,9 @@ package language.objects
 
 import language.DataType
 
-class World(name: String, mutable: Boolean = false): VulcanObject(DataType.WORLD, name, mutable) {
+class World(name: String, java: String = name, mutable: Boolean = false): VulcanObject(DataType.WORLD, name, java, mutable) {
 
-    override val validMessages: Map<String, Int> = mapOf(
+    override val actions: Map<String, Int> = mapOf(
             Pair("sunshine", 3),
             Pair("rain", 3),
             Pair("thunder", 3),
@@ -16,11 +16,11 @@ class World(name: String, mutable: Boolean = false): VulcanObject(DataType.WORLD
             "sunshine" -> {
                 if(parameters[0] == "for" && (parameters[2] == "seconds" || parameters[2] == "second")) {
                     val time = DataType.INTEGER.toJava(parameters[1], variables)
-                    return "$name.getWorldInfo().setCleanWeatherTime($time * 20);¶" +
-                            "$name.getWorldInfo().setRainTime(0);¶" +
-                            "$name.getWorldInfo().setThunderTime(0);¶" +
-                            "$name.getWorldInfo().setRaining(false);¶" +
-                            "$name.getWorldInfo().setThundering(false);"
+                    return "$java.getWorldInfo().setCleanWeatherTime($time * 20);¶" +
+                            "$java.getWorldInfo().setRainTime(0);¶" +
+                            "$java.getWorldInfo().setThunderTime(0);¶" +
+                            "$java.getWorldInfo().setRaining(false);¶" +
+                            "$java.getWorldInfo().setThundering(false);"
                 } else {
                     throw IllegalArgumentException("invalid syntax")
                 }
@@ -29,11 +29,11 @@ class World(name: String, mutable: Boolean = false): VulcanObject(DataType.WORLD
             "rain" -> {
                 if(parameters[0] == "for" && (parameters[2] == "seconds" || parameters[2] == "second")) {
                     val time = DataType.INTEGER.toJava(parameters[1], variables)
-                    return "$name.getWorldInfo().setCleanWeatherTime(0);¶" +
-                            "$name.getWorldInfo().setRainTime($time * 20);¶" +
-                            "$name.getWorldInfo().setThunderTime(0);¶" +
-                            "$name.getWorldInfo().setRaining(true);¶" +
-                            "$name.getWorldInfo().setThundering(false);"
+                    return "$java.getWorldInfo().setCleanWeatherTime(0);¶" +
+                            "$java.getWorldInfo().setRainTime($time * 20);¶" +
+                            "$java.getWorldInfo().setThunderTime(0);¶" +
+                            "$java.getWorldInfo().setRaining(true);¶" +
+                            "$java.getWorldInfo().setThundering(false);"
                 } else {
                     throw IllegalArgumentException("invalid syntax")
                 }
@@ -42,11 +42,11 @@ class World(name: String, mutable: Boolean = false): VulcanObject(DataType.WORLD
             "thunder" -> {
                 if(parameters[0] == "for" && (parameters[2] == "seconds" || parameters[2] == "second")) {
                     val time = DataType.INTEGER.toJava(parameters[1], variables)
-                    return "$name.getWorldInfo().setCleanWeatherTime(0);¶" +
-                            "$name.getWorldInfo().setRainTime($time * 20);¶" +
-                            "$name.getWorldInfo().setThunderTime($time * 20);¶" +
-                            "$name.getWorldInfo().setRaining(true);¶" +
-                            "$name.getWorldInfo().setThundering(true);"
+                    return "$java.getWorldInfo().setCleanWeatherTime(0);¶" +
+                            "$java.getWorldInfo().setRainTime($time * 20);¶" +
+                            "$java.getWorldInfo().setThunderTime($time * 20);¶" +
+                            "$java.getWorldInfo().setRaining(true);¶" +
+                            "$java.getWorldInfo().setThundering(true);"
                 } else {
                     throw IllegalArgumentException("invalid syntax")
                 }
@@ -56,7 +56,7 @@ class World(name: String, mutable: Boolean = false): VulcanObject(DataType.WORLD
                 if(parameters[0] == "with" && parameters[1] == "strength" && parameters[3] == "at") {
                     val strength = DataType.FLOAT.toJava(parameters[2], variables)
                     val position = DataType.VECTOR3.toJava(parameters[4], variables)
-                    return "$name.createExplosion(null, $position.getX(), $position.getY(), $position.getZ(), $strength, true);"
+                    return "$java.createExplosion(null, $position.getX(), $position.getY(), $position.getZ(), $strength, true);"
                 } else {
                     throw IllegalArgumentException("invalid syntax")
                 }