+++ Changelog.md
@@ -7,6 +7,7 @@ In development
 * Added several new data types: boolean, integer, decimal, string, vector and world (note that most of these do not yet have all of the desired features, such as an easy way to perform arithmetic)
 * All block behaviours now have access to the world
 * Added an attribute class to store the value and data type of the different attributes for items, blocks, etc
+* Lines starting with invalid words will now throw an error
 
 ## Alpha 0.2.0
 Not yet publicly released due to lack of internet connection